requirements.txt
click==6.7
click-plugins==1.0.3
cligj==0.4.0
munch==2.1.1
olefile==0.44
Pillow==4.1.1
Shapely==1.5.15
six==1.10.0
